set default collapsed
