+++ src/panorama/react/hud_lottery/KeyBind.tsx
@@ -1,10 +1,11 @@
 import 'panorama-polyfill-x/lib/console';
 import 'panorama-polyfill-x/lib/timers';
 
-import { useState } from 'react';
+import { useState, useEffect } from 'react';
 import ExpandButton from './components/ExpandButton';
 import KeyBindContainer from './components/KeyBindContainer';
 import { GetLocalPlayerSteamAccountID } from '@utils/utils';
+import { GetPlayer } from '@utils/net-table';
 
 const containerStyle: Partial<VCSSStyleDeclaration> = {
   flowChildren: 'down',
@@ -19,6 +20,21 @@ function KeyBind() {
     setIsCollapsed(!isCollapsed);
   };
 
+  const playerSetting = GetPlayer(GetLocalPlayerSteamAccountID())?.playerSetting ?? {
+    isRememberAbilityKey: true,
+    activeAbilityKey: '',
+    passiveAbilityKey: '',
+    activeAbilityQuickCast: false,
+    passiveAbilityQuickCast: false,
+  };
+
+  useEffect(() => {
+    // 如果快捷键有设置，则默认折叠
+    if (playerSetting.activeAbilityKey || playerSetting.passiveAbilityKey) {
+      setIsCollapsed(true);
+    }
+  }, [playerSetting.activeAbilityKey, playerSetting.passiveAbilityKey]);
+
   // 获取玩家steamId，如果获取失败，则为观战，不显示
   const steamAccountId = GetLocalPlayerSteamAccountID();
   if (!steamAccountId) {
@@ -27,7 +43,7 @@ function KeyBind() {
 
   return (
     <Panel style={containerStyle}>
-      <KeyBindContainer isCollapsed={isCollapsed} />
+      <KeyBindContainer isCollapsed={isCollapsed} playerSetting={playerSetting} />
       <Panel style={{ horizontalAlign: 'left', verticalAlign: 'bottom' }}>
         <ExpandButton
           textToken={isCollapsed ? '#key_bind' : '#key_bind_collapsed'}
