+++ src/panorama/react/hud_lottery/components/KeyBindContainer.tsx
@@ -1,24 +1,18 @@
 import React, { useEffect, useState } from 'react';
 import KeySettingButton from './KeySettingButton';
 import KeyBindRemember from './KeyBindRemember';
-import { GetLotteryStatus, GetPlayer, SubscribeLotteryStatus } from '@utils/net-table';
+import { GetLotteryStatus, SubscribeLotteryStatus } from '@utils/net-table';
 import { GetLocalPlayerSteamAccountID } from '@utils/utils';
 import { LotteryStatusDto } from '../../../../common/dto/lottery-status';
 import { saveInputKeyborard } from '../hotkey';
+import { PlayerSetting } from '../../../../vscripts/api/player';
 
 interface KeyBindContainerProps {
   isCollapsed: boolean;
+  playerSetting: PlayerSetting;
 }
 
-const KeyBindContainer: React.FC<KeyBindContainerProps> = ({ isCollapsed }) => {
-  const playerSetting = GetPlayer(GetLocalPlayerSteamAccountID())?.playerSetting ?? {
-    isRememberAbilityKey: true,
-    activeAbilityKey: '',
-    passiveAbilityKey: '',
-    activeAbilityQuickCast: false,
-    passiveAbilityQuickCast: false,
-  };
-
+const KeyBindContainer: React.FC<KeyBindContainerProps> = ({ isCollapsed, playerSetting }) => {
   const containerStyle: Partial<VCSSStyleDeclaration> = {
     visibility: isCollapsed ? 'collapse' : 'visible',
     flowChildren: 'down',
@@ -49,6 +43,14 @@ const KeyBindContainer: React.FC<KeyBindContainerProps> = ({ isCollapsed }) => {
   }, [steamAccountId]);
 
   useEffect(() => {
+    // 发送快捷键设置到服务器端进行保存
+    GameEvents.SendCustomGameEventToServer('save_bind_ability_key', {
+      isRememberAbilityKey,
+      activeAbilityKey,
+      passiveAbilityKey,
+      activeAbilityQuickCast,
+      passiveAbilityQuickCast,
+    });
     // 每秒刷新一次改键显示
     const timer = setInterval(() => {
       saveInputKeyborard(
@@ -61,18 +63,8 @@ const KeyBindContainer: React.FC<KeyBindContainerProps> = ({ isCollapsed }) => {
     return () => {
       clearInterval(timer);
     };
-  }, [lotteryStatus, activeAbilityKey, passiveAbilityKey, isRememberAbilityKey]);
-
-  // 发送快捷键设置到服务器端进行保存
-  useEffect(() => {
-    GameEvents.SendCustomGameEventToServer('save_bind_ability_key', {
-      isRememberAbilityKey,
-      activeAbilityKey,
-      passiveAbilityKey,
-      activeAbilityQuickCast,
-      passiveAbilityQuickCast,
-    });
   }, [
+    lotteryStatus,
     isRememberAbilityKey,
     activeAbilityKey,
     passiveAbilityKey,
